+++ src/test/java/org/morphix/lang/NullablesTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2025 the original author or authors.
+ * Copyright 2026 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
  * the License. You may obtain a copy of the License at
@@ -155,6 +155,18 @@ class NullablesTest {
 				.andNotNull(A::getB)
 				.andNotNull(B::getS)
 				.valueOrDefault(BIBI);
+
+		assertThat(result, equalTo(BIBI));
+	}
+
+	@Test
+	void shouldReturnDefaultOnNestingAndNotNullEvenIfOnePathIsNullAndOrElse() {
+		A a = new A();
+
+		String result = Nullables.whenNotNull(a)
+				.andNotNull(A::getB)
+				.andNotNull(B::getS)
+				.orElse(BIBI);
 
 		assertThat(result, equalTo(BIBI));
 	}