+++ src/test/java/org/morphix/reflection/FieldsGetOneDeclaredInHierarchyTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2025 the original author or authors.
+ * Copyright 2026 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
  * the License. You may obtain a copy of the License at
@@ -20,7 +20,12 @@ import java.lang.reflect.Field;
 import org.junit.jupiter.api.Test;
 
 /**
- * Test class for {@link Fields#getOneDeclaredInHierarchy(Class, String)}.
+ * Test class for:
+ *
+ * <ul>
+ * <li>{@link Fields#getOneDeclaredInHierarchy(Class, String)}</li>
+ * <li>{@link Fields#getOneDeclaredInHierarchy(Object, String)}</li>
+ * </ul>
  *
  * @author Radu Sebastian LAZIN
  */
@@ -64,6 +69,13 @@ class FieldsGetOneDeclaredInHierarchyTest {
 	@Test
 	void shouldReturnNullIfClassIsNull() {
 		Field field = Fields.getOneDeclaredInHierarchy(null, "x");
+
+		assertThat(field, equalTo(null));
+	}
+
+	@Test
+	void shouldReturnNullIfObjectIsNull() {
+		Field field = Fields.getOneDeclaredInHierarchy((Object) null, "x");
 
 		assertThat(field, equalTo(null));
 	}